+++ BaseDatosTPC/Ticket.cs
@@ -18,6 +18,7 @@ namespace BaseDatosTPC
 
         public string? ID_Proveedor { get; set; }
 
+
         public string? Detalle { get; set; }
 
         public int? Solped { get; set; }
@@ -25,7 +26,7 @@ namespace BaseDatosTPC
         public string? Id_OE { get; set; }
         public int N_OE { get; set; }
       
-        public int Numero_OC { get; set; }
+        public long Numero_OC { get; set; }
         public DateTime? Fecha_OC_Recepcionada { get; set; }
 
         public DateTime? Fecha_OC_Enviada { get; set; }
